Refactor comment sorting order mechanism
This commit updates the comment sorting order in the background.js file to descending order. It also changes the list layout in the popup.css file to display items in a column instead of a reverse column.

Instead of sorting with CSS, we'll now sort with the request parameter.

This commit also fixes a bug with sorting in the configurable summary output since the data itself is now sorted.